requirements.txt
python-dotenv
pymysql
requests
termcolor
sentence_transformers
numpy
scikit-learn
pandas
tqdm
ipykernel